"fix? AI did it. Review"

+++ pkg/serror/internal/record/value.go
@@ -250,11 +250,11 @@ func AnyValue(v any) Value {
 	case uintptr:
 		return Uint64Value(uint64(v))
 	case int8:
-		return IntValue(int(v))
+		return Int64Value(int64(v))
 	case int16:
-		return IntValue(int(v))
+		return Int64Value(int64(v))
 	case int32:
-		return IntValue(int(v))
+		return Int64Value(int64(v))
 	case float64:
 		return Float64Value(v)
 	case float32:
